--- sky/dashboard/src/pages/gpus.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import { GPUs } from '@/components/gpus';
-import Head from 'next/head';
-
-export default function GPUsPage() {
-  return (
-    <>
-      <Head>
-        <title>GPUs | SkyPilot Dashboard</title>
-      </Head>
-      <GPUs />
-    </>
-  );
-}
