+++ sky/dashboard/src/pages/infra.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import { GPUs } from '@/components/infra';
+import Head from 'next/head';
+
+export default function InfraPage() {
+  return (
+    <>
+      <Head>
+        <title>Infra | SkyPilot Dashboard</title>
+      </Head>
+      <GPUs />
+    </>
+  );
+}
